+++ src/connectorgen/utils/version.go
@@ -1,0 +1,10 @@
+package utils
+
+type Version struct {
+	CurrentVersion string `json:"currentVersion"`
+	UsingLatest    bool   `json:"usingLatest"`
+}
+
+func CompareVersions(repoVersion, latestVersion string) bool {
+	return repoVersion == latestVersion
+}
